added some content

+++ src/com/company/engine/math/Point.java
@@ -4,7 +4,7 @@ public class Point {
     private float x;
     private float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
